streamlit-app/requirements.txt
streamlit
pandas
pandas_gbq
numpy
google-auth
streamlit-autorefresh